Allow multiple providers for vmware transport
Currently, the transport resource does not have any
providers, it is intended as a simple way to specify
data as resources.

This patch adds basic support for multiple providers so
that credentials can be passed in ways other than as
resources.

This is done by calling the getter methods for
the properties that are used to retrieve credential
information.

For my use case, it will allow me to store credentials in
device.conf files (and to use my custom encryption/decryption
mechanisms.